Move remaining metadata functionality out of util
