+++ gslib/command.py
@@ -2080,7 +2080,7 @@ class WorkerThread(threading.Thread):
     self.user_project = user_project
 
     # Note that thread_gsutil_api is not initialized in the sequential
-    # case; task functions should use util.GetCloudApiInstance to
+    # case; task functions should use cloud_api.GetCloudApiInstance to
     # retrieve the main thread's CloudApiDelegator in that case.
     self.thread_gsutil_api = None
     if bucket_storage_uri_class and gsutil_api_map:
