+++ gslib/commands/setmeta.py
@@ -20,6 +20,7 @@ import time
 
 from apitools.base.py import encoding
 from gslib.cloud_api import AccessDeniedException
+from gslib.cloud_api import GetCloudApiInstance
 from gslib.cloud_api import PreconditionException
 from gslib.cloud_api import Preconditions
 from gslib.command import Command
@@ -35,9 +36,8 @@ from gslib.thread_message import MetadataMessage
 from gslib.translation_helper import CopyObjectMetadata
 from gslib.translation_helper import ObjectMetadataFromHeaders
 from gslib.translation_helper import PreconditionsFromHeaders
-from gslib.util import GetCloudApiInstance
-from gslib.util import IsCustomMetadataHeader
 from gslib.utils.constants import NO_MAX
+from gslib.utils.metadata_util import IsCustomMetadataHeader
 from gslib.utils.retry_util import Retry
 from gslib.utils.text_util import InsistAsciiHeader
 from gslib.utils.text_util import InsistAsciiHeaderValue
