+++ gslib/posix_util.py
@@ -24,8 +24,8 @@ import time
 
 from gslib.exception import CommandException
 from gslib.tz_utc import UTC
-from gslib.util import CreateCustomMetadata
-from gslib.util import GetValueFromObjectCustomMetadata
+from gslib.utils.metadata_util import CreateCustomMetadata
+from gslib.utils.metadata_util import GetValueFromObjectCustomMetadata
 from gslib.utils.system_util import IS_WINDOWS
 from gslib.utils.unit_util import SECONDS_PER_DAY
