+++ gslib/utils/update_util.py
@@ -14,6 +14,9 @@
 # limitations under the License.
 """Shared utility methods for the update command and its tests."""
 
+from __future__ import absolute_import
+from __future__ import print_function
+
 import logging
 import os
 import re
